Show device icon when player is off
