+++ src/components/MediocreMediaPlayerCard/components/AlbumArt.tsx
@@ -1,16 +1,18 @@
 import styled from "@emotion/styled";
 import { Icon, usePlayer } from "@components";
 import { ButtonHTMLAttributes, Fragment, JSX } from "preact/compat";
+import { getDeviceIcon } from "@utils";
 
 const AlbumArtContainer = styled.button<{
-  maxHeight: string;
+  $state: string;
 }>`
   background: none;
   border: none;
   cursor: pointer;
   padding: 0px;
   margin: 0px;
-  max-height: ${props => props.maxHeight};
+  max-height: ${props => (props.$state === "off" ? "80px" : "100px")};
+  height: ${props => (props.$state === "off" ? "80px" : "100px")};
   aspect-ratio: 1;
   flex-shrink: 0;
   border-radius: 4px;
@@ -30,6 +32,7 @@ const NoAlbumArt = styled.div`
   align-items: center;
   width: 100%;
   height: 100%;
+  aspect-ratio: 1;
   color: var(--card-background-color);
   background-color: var(--primary-text-color);
   opacity: 0.5;
@@ -44,13 +47,11 @@ const SourceIndicator = styled.div`
 `;
 
 export type AlbumArtProps = {
-  maxHeight: string;
   renderLongPressIndicator?: () => JSX.Element | null;
 } & ButtonHTMLAttributes<HTMLButtonElement>;
 
 // Component
 export const AlbumArt = ({
-  maxHeight,
   renderLongPressIndicator,
   ...buttonProps
 }: AlbumArtProps) => {
@@ -59,22 +60,31 @@ export const AlbumArt = ({
     media_title: title,
     media_artist: artist,
     entity_picture: albumArt,
+    icon,
+    device_class: deviceClass,
     source,
   } = player.attributes;
   const state = player.state;
 
   return (
-    <AlbumArtContainer maxHeight={maxHeight} {...buttonProps}>
+    <AlbumArtContainer $state={state} {...buttonProps}>
       {albumArt ? (
         <Fragment>
           <AlbumArtImage src={albumArt} alt={`${title} by ${artist}`} />
           <SourceIndicator>
-            <Icon size="xx-small" icon={getIcon({ source, state })} />
+            <Icon size="xx-small" icon={getIcon({ source })} />
           </SourceIndicator>
         </Fragment>
       ) : (
         <NoAlbumArt>
-          <Icon size="x-large" icon={getIcon({ source, state })} />
+          <Icon
+            size="large"
+            icon={
+              state === "off"
+                ? getDeviceIcon({ icon, deviceClass })
+                : getIcon({ source })
+            }
+          />
         </NoAlbumArt>
       )}
       {!!renderLongPressIndicator && renderLongPressIndicator()}
@@ -82,8 +92,7 @@ export const AlbumArt = ({
   );
 };
 
-const getIcon = ({ source, state }: { source: string; state: string }) => {
-  if (state === "off") return "mdi:power-off";
+const getIcon = ({ source }: { source: string }) => {
   switch (source?.toLowerCase()) {
     case "spotify":
       return "mdi:spotify";
