+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./actions";
+export * from "./getDeviceIcon";
 export * from "./getDidMediaPlayerUpdate";
 export * from "./getHass";
